Preload video if videoId is given
This patch speeds up loading the first video that's played with a
react-youtube instance. Previously, the YT player would be created
and had to be loaded completely before a video was loaded in using
loadVideoById. With this patch, the YouTube API creates an iframe
with the video details already in it, allowing it to start playing
the video one or two seconds earlier.

To illustrate the difference, an example iframe src URL before the
patch:

    https://www.youtube.com/embed/?autoplay=1&...

and after the patch:

    https://www.youtube.com/embed/2gWIjRqvtnI?autoplay=1&...

When a new video is loaded later because of a change in
`props.videoId`, that video will just play in the same iframe too,
so that stays just as speedy as it was before.