Do not reset player when playerVars.{start,end} change
A change in the start or end time of a video can be resolved by
calling `updateVideo`, instead of resetting the player. This makes
loading the next video a bit faster if it has a different start or
end time than the previous video. If the next video is loaded when
the user is not on the tab, this also helps deal with Chrome's
[gesture requirement for media playback](http://techdows.com/2015/08/future-chrome-versions-wont-autoplay-media-in-the-background-tabs.html),
by reusing the same YouTube API iframe instead of creating a new
one.

Now a player reset is triggered if `props.opts` changes, ignoring
`playerVars.start` and `playerVars.end`. All other `props.opts`
properties still trigger a full reset.

A video update is triggered when `props.videoId` changes, but also
when `playerVars.start` or `playerVars.end` change.

In the common case, `playerVars.{start,end}` will only change when
`props.videoId` changes, and that will now only trigger a video
update instead of a player reset :)

A future optimisation could be to use YouTube's `seekTo` method if
`playerVars.start` is the only changed property.