prevent YT API from replacing React nodes, fixes #74
When React tries to unmount a component, it expects its own DOM
nodes to exist, so it can remove them. However, the YouTube API
_replaces_ the node with the ID you give it. So, when React tries
to unmount a react-youtube component, it tries to remove a DOM node
that was replaced by a different DOM node and no longer exists.

React-YouTube _does_ call destroy() on the YouTube API, but because
the youtube-player wraps it in a Promise resolution handler, it
gets executed _after_ React has removed DOM nodes.

This patch works around that by wrapping the DOM node that gets
replaced by the YouTube API, in another DOM node that is left
untouched. That way, React can safely remove the wrapping DOM
node. The YouTube destroy() call happens a tick later, but that
is just fine.

As a side effect of this change, the component now passes the
actual DOM element to `youtube-player`, instead of a generated
ID.

This changes the tests to account for the `id` and `className`
properties now no longer applying to the root element, and for
the new way of instantiating the YouTube API.